requirements.txt
# ── Core ──────────────────────────────────────────────────────────────
ccxt>=4.2.0
pandas>=2.0.0
numpy>=1.24.0
ta>=0.10.2
scikit-learn>=1.3.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
schedule>=1.2.0
requests>=2.31.0
joblib>=1.3.0
aiohttp>=3.9.0

# ── ML ensemble + tuning ──────────────────────────────────────────────
lightgbm>=4.0.0
xgboost>=2.0.0
catboost>=1.2.0
optuna>=3.4.0

# ── Deep learning ─────────────────────────────────────────────────────
torch>=2.1.0

# ── Dashboard ─────────────────────────────────────────────────────────
streamlit>=1.28.0
plotly>=5.17.0

# ── RL (legacy) ───────────────────────────────────────────────────────
stable-baselines3[extra]>=2.3.0
gymnasium>=0.29.1

# ── Solana spot ───────────────────────────────────────────────────────
solana>=0.34.0
solders>=0.21.0
base58>=2.1.1

# ── Hyperliquid perp ──────────────────────────────────────────────────
hyperliquid-python-sdk>=0.4.0
eth-account>=0.11.0

# ── Telegram real-time news ───────────────────────────────────────────
telethon>=1.34.0

# ── API server ────────────────────────────────────────────────────────
flask>=3.0.0
